rename transformer to factoryConstructorTransformer
